+++ controllers/gamesController.js
@@ -1,12 +1,16 @@
 const Game = require('../models/Game') 
 const { playerTypes, gameGenres, gamePlatforms } = require('../assets/enums') 
 const Developer = require('../models/Developer')
+const mongoose = require('mongoose')
+
 
 exports.getGamesByDeveloper = async (req, res) => {
   try {
     const { expand } = req.query 
+    if (!mongoose.Types.ObjectId.isValid(req.params.developerId)) {
+      return res.status(404).json({ error: 'Developer not found' })
+    }
     let gamesQuery = Game.find({ developerId: req.params.developerId }) 
-    console.log(req.params.developerId)
 
     if (expand && expand.includes('developerId')) {
       gamesQuery = gamesQuery.populate('developerId') 
@@ -16,8 +20,13 @@ exports.getGamesByDeveloper = async (req, res) => {
       gamesQuery = gamesQuery.populate('userId') 
     }
 
-    const games = await gamesQuery 
-    res.json(games) 
+    const games = await gamesQuery
+    if(games){
+      res.json(games)
+    } else{
+      return res.status(404).json({ error: 'Game not found' })
+    }
+ 
   } catch (error) {
     res.status(500).json({ error: 'Server error' }) 
   }
@@ -35,7 +44,11 @@ exports.getAllGames = async (req, res) => {
     }
 
     const games = await gamesQuery
-    res.json(games) 
+    if(games){
+      res.json(games)
+    } else{
+      return res.status(404).json({ error: 'Game not found' })
+    }
   } catch (error) {
     res.status(500).json({ error: 'Server error' }) 
   }
@@ -45,18 +58,19 @@ exports.getAllGames = async (req, res) => {
 exports.getGameById = async (req, res) => {
   try {
     const { expand } = req.query 
-    let gameQuery = Game.findById(req.params.id) 
 
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      return res.status(404).json({ error: 'Game not found' })
+    }
+
+    let gameQuery = Game.findById(req.params.id) 
     if(expand && expand.includes('developerId')){
       gameQuery = gameQuery.populate('developerId')
     }
 
     const game = await gameQuery 
-    if (game) {
-      res.json(game) 
-    } else {
-      res.status(404).json({ error: 'Game not found' }) 
-    }
+    res.json(game) 
+
   } catch (error) {
     res.status(500).json({ error: 'Server error' }) 
   }
@@ -70,11 +84,15 @@ exports.createGame = async (req, res) => {
       language, playerType, developerId
     } = req.body 
 
+
+    if (!mongoose.Types.ObjectId.isValid(developerId)) {
+      return res.status(404).json({ error: 'Developer not found' })
+    }
     if (!Object.values(playerTypes).includes(playerType)) {
-      return res.status(400).json({ error: 'Invalid player type.' }) 
+      return res.status(422).json({ error: 'Invalid player type.' }) 
     }
     if (!Object.values(gameGenres).includes(genre)) {
-      return res.status(400).json({ error: 'Invalid game genre.' }) 
+      return res.status(422).json({ error: 'Invalid game genre.' }) 
     }
 
     const developer = await Developer.findById(developerId) 
@@ -95,8 +113,7 @@ exports.createGame = async (req, res) => {
     await newGame.save() 
     res.status(201).json(newGame) 
   } catch (error) {
-    console.log(error)
-    res.status(400).json({ error: 'Bad request' }) 
+    res.status(500).json({ error: 'Error creating game' })
   }
 } 
 
@@ -107,14 +124,21 @@ exports.updateGame = async (req, res) => {
       title, genre, platform, controllerSupport,
       language, playerType, developerId
     } = req.body 
-
-
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      return res.status(404).json({ error: 'Game not found' })
+    }
+    if (!mongoose.Types.ObjectId.isValid(developerId)) {
+      return res.status(404).json({ error: 'Developer not found' })
+    }
+    if(!title || !genre || !platform || !controllerSupport || !language || !playerType || !developerId){
+      return res.status(400).json({ error: 'Need all fields' }) 
+    }
     if (playerType && !Object.values(playerTypes).includes(playerType)) {
-      return res.status(400).json({ error: 'Invalid player type.' }) 
+      return res.status(422).json({ error: 'Invalid player type.' }) 
     }
  
     if (genre && !Object.values(gameGenres).includes(genre)) {
-      return res.status(400).json({ error: 'Invalid game genre.' }) 
+      return res.status(422).json({ error: 'Invalid game genre.' }) 
     }
     if(developerId){
       const developer = await Developer.findById(developerId) 
@@ -136,13 +160,16 @@ exports.updateGame = async (req, res) => {
       res.status(404).json({ error: 'Game not found' }) 
     }
   } catch (error) {
-    res.status(400).json({ error: 'Bad request' }) 
+    res.status(500).json({ error: 'Error updating game' })
   }
 } 
 
 // DELETE a game
 exports.deleteGame = async (req, res) => {
   try {
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      return res.status(404).json({ error: 'Game not found' })
+    }
     const deletedGame = await Game.findByIdAndDelete(req.params.id) 
     if (deletedGame) {
       res.json(deletedGame)
