+++ controllers/usersController.js
@@ -1,5 +1,8 @@
+const Developer = require('../models/Developer')
 const User = require('../models/User') 
 const bcrypt = require('bcrypt') 
+const mongoose = require('mongoose')
+
 
 const isValidEmail = (email) => {
   const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/ 
@@ -19,26 +22,48 @@ exports.getAllUsers = async (req, res) => {
 // GET a single user by ID
 exports.getUserById = async (req, res) => {
   try {
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      return res.status(404).json({ error: 'User not found' })
+    }
     const user = await User.findById(req.params.id)
+    
     if (user) {
       res.json(user)
     } else {
       res.status(404).json({ error: 'User not found' })
     }
   } catch (error) {
-    res.status(500).json({ error: 'Error fetching user' })
+    res.status(500).json({ error: 'Error getting user' })
+  }
+} 
+
+exports.getUsersDevelopers = async (req, res) => {
+  try {
+    if (!mongoose.Types.ObjectId.isValid(req.params.userId)) {
+      return res.status(404).json({ error: 'User not found' })
+    }
+    let userQuery = Developer.find({ userId: req.params.userId })
+
+    const user = await userQuery
+    if (user) {
+      res.json(user)
+    } else {
+      res.status(404).json({ error: 'User not found' })
+    }
+  } catch (error) {
+    res.status(500).json({ error: 'Error getting user developers' })
   }
 } 
 
 // POST a new user
 exports.createUser = async (req, res) => {
   try {
-    const { username, email, password, type } = req.body 
 
+    const { username, email, password, type } = req.body 
 
     const validTypes = ['admin', 'dev', 'guest'] 
     if (!validTypes.includes(type)) {
-      return res.status(400).json({ error: 'Invalid user type. Valid types are: admin, dev, guest' }) 
+      return res.status(422).json({ error: 'Invalid user type. Valid types are: admin, dev, guest' }) 
     }
 
     if (!isValidEmail(email)) {
@@ -57,7 +82,7 @@ exports.createUser = async (req, res) => {
     res.status(201).json(savedUser) 
   } catch (error) {
     if (error.code === 11000) {
-      res.status(400).json({ error: 'Email already exists' }) 
+      res.status(422).json({ error: 'Email already exists' }) 
     } else {
       res.status(500).json({ error: 'Error creating user' }) 
     }
@@ -68,21 +93,26 @@ exports.createUser = async (req, res) => {
 exports.updateUser = async (req, res) => {
   try {
     const { username, email, password, type } = req.body 
-
+    if(!username || !email || !password || !type){
+      return res.status(400).json({ error: 'Need all fields' }) 
+    }
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      return res.status(404).json({ error: 'User not found' })
+    }
     const validTypes = ['admin', 'dev', 'guest'] 
     if (!validTypes.includes(type)) {
-      return res.status(400).json({ error: 'Invalid user type. Valid types are: admin, dev, guest' }) 
+      return res.status(422).json({ error: 'Invalid user type. Valid types are: admin, dev, guest' }) 
     }
 
     if (email && !isValidEmail(email)) {
-      return res.status(400).json({ error: 'Invalid email format' }) 
+      return res.status(422).json({ error: 'Invalid email format' }) 
     }
 
     let updatedData = { username, email, type } 
-    if (password) {
-      const salt = await bcrypt.genSalt(10) 
-      updatedData.password = await bcrypt.hash(password, salt) 
-    }
+
+    const salt = await bcrypt.genSalt(10) 
+    updatedData.password = await bcrypt.hash(password, salt) 
+    
 
     const updatedUser = await User.findByIdAndUpdate(
       req.params.id,
@@ -103,6 +133,9 @@ exports.updateUser = async (req, res) => {
 // DELETE a user
 exports.deleteUser = async (req, res) => {
   try {
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      return res.status(404).json({ error: 'User not found' })
+    }
     const deletedUser = await User.findByIdAndDelete(req.params.id)  
     if (deletedUser) {
       res.json(deletedUser)
